v2.0

+++ routes/api/categoryAPI.js
@@ -4,6 +4,7 @@ const {check, validationResult} = require('express-validator');
 
 const auth = require('../../middleware/auth');
 const Category = require('../../models/Category');
+const Product = require('../../models/Product');
 
 const categoryValidator = [
   //name
@@ -31,7 +32,6 @@ module.exports = function categoryAPI(router) {
         if (!errors.isEmpty()) {
           return res.status(422).json({errors: errors.array()});
         }
-
         const {name} = req.body;
         const category = await Category.findOne({name});
         if (category) {
@@ -48,10 +48,18 @@ module.exports = function categoryAPI(router) {
 //DELETE - delete category by id
   router.delete('/categories', async (req, res) => {
     try {
+
       const categoryId = req.body.id;
       let category = await Category.findById(categoryId);
       if (!category) {
         return res.status(404).send('Category does not exist');
+      }
+      //check if any product uses this id
+      const products = await Product.find();
+      //Cannot use === here, item.category is object, categoryId is string
+      let hasId = products.some(item => item.category == categoryId);
+      if (hasId) {
+        return res.status(400).send('Sorry, A Product use this Id, cannot delete');
       }
       await Category.findByIdAndRemove({_id: categoryId});
       res.json({msg: 'Category deleted'});
